# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Ingestor System API
  description: |-
    RESTful API for the Ingestor System that provides access to content processing,
    entity extraction, and database capabilities using Claude AI.
  version: 1.0.0
  contact:
    email: support@example.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'http://localhost:3000/api/v1'
    description: Local development server
...
paths:
...
  /auth/me:
    get:
      tags:
        - auth
      summary: Get current user info
      description: Returns information about the authenticated user
      operationId: getCurrentUser
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
  /auth/change-password:
    post:
      tags:
        - auth
      summary: Change user password
      description: Changes the password for the authenticated user
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - currentPassword
                - newPassword
              properties:
                currentPassword:
                  type: string
                  example: oldpassword
                newPassword:
                  type: string
                  example: newpassword
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: Password changed successfully
        '401':
          description: Current password is incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
  /auth/api-keys:
    post:
      tags:
        - auth
      summary: Generate API key
      description: Generates a new API key for the authenticated user
      operationId: generateApiKey
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                  example: My API Key
      responses:
        '201':
          description: API key generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
    
    get:
      tags:
        - auth
      summary: List API keys
      description: Lists all API keys for the authenticated user
      operationId: listApiKeys
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeysListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
  /auth/api-keys/{id}:
    delete:
      tags:
        - auth
      summary: Delete API key
      description: Deletes an API key
      operationId: deleteApiKey
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: API key deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: API key deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: API key not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
  
  #
  # Content Endpoints
  #
...
components:
  schemas:
...
    UserResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
              example: 123e4567-e89b-12d3-a456-426614174000
            username:
              type: string
              example: admin
            role:
              type: string
              enum: [user, admin]
              example: admin
            createdAt:
              type: string
              format: date-time
              example: '2025-05-14T04:21:56Z'
    ApiKeyResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
              example: 123e4567-e89b-12d3-a456-426614174000
            apiKey:
              type: string
              example: MTIzZTQ1NjctZTg5Yi0xMmQzLWE0NTYtNDI2NjE0MTc0MDAwMTIzZTQ1NjctZTg5Yi0xMmQzLWE0NTYtNDI2NjE0MTc0MDAw
            description:
              type: string
              example: My API Key
            createdAt:
              type: string
              format: date-time
              example: '2025-05-14T04:21:56Z'
    ApiKeysListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
                example: 123e4567-e89b-12d3-a456-426614174000
              description:
                type: string
                example: My API Key
              createdAt:
                type: string
                format: date-time
                example: '2025-05-14T04:21:56Z'
              lastUsed:
                type: string
                format: date-time
                example: '2025-05-14T08:45:33Z'
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Validation error
            details:
              type: array
              items:
                type: object
                properties:
                  path:
                    type: string
                    example: content
                  message:
                    type: string
                    example: Content is required
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key